feat: handle errors

+++ src/handlers/mv.js
@@ -6,26 +6,38 @@ export async function handleMoveFile(source, newDirectory) {
   const sourcePath = `${process.cwd()}/${source}`;
   const destinationPath = path.join(process.cwd(), newDirectory, path.basename(source));
 
-  const readable = fs.createReadStream(sourcePath);
-  const writable = fs.createWriteStream(destinationPath);
+  try {
+    const sourceExists = await fs.promises.access(sourcePath, fs.constants.F_OK)
+      .then(() => true)
+      .catch(() => false);
 
-  await new Promise((resolve, reject) => {
-    readable.pipe(writable);
+    if (!sourceExists) {
+      console.error('Operation failed');
+      return;
+    }
 
-    writable.on('finish', () => {
-      fs.unlink(sourcePath, (error) => {
-        if (error) {
-          reject(error);
-        } else {
-          resolve();
-        }
+    await new Promise((resolve, reject) => {
+      const readable = fs.createReadStream(sourcePath);
+      const writable = fs.createWriteStream(destinationPath);
+
+      readable.pipe(writable);
+
+      writable.on('finish', () => {
+        fs.unlink(sourcePath, (error) => {
+          if (error) {
+            reject(error);
+          } else {
+            resolve();
+          }
+        });
+      });
+
+      writable.on('error', (error) => {
+        reject(error);
       });
     });
-
-    writable.on('error', (error) => {
-      reject(error);
-    });
-  });
-
+  } catch (error) {
+    console.error('Operation failed');
+  }
   getCurrentDir();
 }
